% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licor_reader.R
\name{licor_reader}
\alias{licor_reader}
\title{Licor Reader}
\usage{
licor_reader(filepath)
}
\description{
This function is used to extract sample data from the cumbersome
files generated from the licor respiration measuring device.
The file needs to be moved to the computer, opened with Excel,
and saved as a csv file before processing. When taking measurements, remarks should be entered that start
with the word "sample" i.e. "sample345".
}
