% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_me.R
\name{print_me}
\alias{print_me}
\title{print_me}
\usage{
print_me(
  plot_name = NULL,
  filename = NULL,
  width = 1280,
  height = 853,
  pointsize = 20,
  res = 200,
  type = "cairo"
)
}
\description{
This is a function for printing
}
