% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE_2.R
\name{summarySE_2}
\alias{summarySE_2}
\title{SummarySE_2}
\usage{
summarySE_2(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\description{
This function summarizes data by groups
}
